lr1110: fold variant trait into Config
The Lr1110Variant trait carried no chip-specific behavior: every impl just
returned a stored field, and Lr1120/Lr1121 were placeholders. Replace it with
plain fields on Config (pa_selection and dio_as_rf_switch) and drop the trait
and its generic parameter.

PA supply is now derived from the PA selection per the LR1110 User Manual
(Tables 9-1 and 9-2): the LP PA runs from the internal regulator, the HP and
HF PAs from VBAT. This also fixes the old Lr1120/Lr1121 placeholders, which
forgot to override the supply and wrongly used VBAT even for the LP PA.

The default RF switch bitmask moves to a Default impl on SetDioAsRfSwitchParams.

Closes #443
